feat: moves github user to common schemas
