+++ backend/services/gitbridge/src/schemas/common.py
@@ -3,3 +3,24 @@ from pydantic import BaseModel, ConfigDict
 
 class CommonSchema(BaseModel):
     model_config = ConfigDict(frozen=True)
+
+
+class GithubUser(CommonSchema):
+    login: str
+    id: int
+    node_id: str
+    avatar_url: str
+    gravatar_id: str
+    url: str
+    html_url: str
+    followers_url: str
+    following_url: str
+    gists_url: str
+    starred_url: str
+    subscriptions_url: str
+    organizations_url: str
+    repos_url: str
+    events_url: str
+    received_events_url: str
+    type: str
+    site_admin: bool
